[PATCH] Fix xxx_get_biosgeometry --- avoid useless 64-bit division.
From Peter Chubb

The bios geometry is almost useless, except for fdisk to try to write
an MSDOS partition table that is vaguely compatible with one written by
other operating systems.

If the size of disc will overflow a ten-bit cylinder number, then all
bets are off anyway.  So fake it by casting the true disc capacity to a
smaller type (than u64), so that we avoid 64-bit division on 32-bit
platforms.  If the disc is small enough that the number of cylinders is
correct, then this has no effect; otherwise, the number-of-cylinders we
report is bogus, but you can't use an MSDOS-format partition table on
such a drive anyway --- use the EFI GPT or the LDM partitioning, which
use 64-bit offsets internally.

+++ drivers/scsi/aha152x.c
@@ -1838,7 +1838,7 @@ int aha152x_biosparam(Scsi_Disk * disk, struct block_device *bdev, int *info_arr
 	/* try default translation */
 	info_array[0] = 64;
 	info_array[1] = 32;
-	info_array[2] = disk->capacity / (64 * 32);
+	info_array[2] = (unsigned long)disk->capacity / (64 * 32);
 
 	/* for disks >1GB do some guessing */
 	if (info_array[2] >= 1024) {
@@ -1853,7 +1853,7 @@ int aha152x_biosparam(Scsi_Disk * disk, struct block_device *bdev, int *info_arr
 				       "         using extended translation.\n");
 				info_array[0] = 255;
 				info_array[1] = 63;
-				info_array[2] = disk->capacity / (255 * 63);
+				info_array[2] = (unsigned long)disk->capacity / (255 * 63);
 			} else {
 				printk(KERN_NOTICE
 				       "aha152x: unable to verify geometry for disk with >1GB.\n"
